virtual bytes fix addendum

+++ src/WindowsAzureDiskResizer/WindowsAzureDiskResizer/Program.cs
@@ -51,7 +51,7 @@ namespace WindowsAzureDiskResizer
             }
 
             // Verify size. Size should be a whole number when converted to MB -and- produce a valid geometry
-            decimal newSizeInGb = newSize / 1024 / 1024 / 1024;
+            decimal newSizeInGb = (decimal)newSize / 1024 / 1024 / 1024;
             if (Math.Round(newSizeInGb) != newSizeInGb)
             {
                 decimal proposedNewSize = Math.Round(newSizeInGb) * 1024 * 1024 * 1024;
